requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
blinker==1.4
cachelib==0.6.0
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.0
cycler==0.11.0
et-xmlfile==1.1.0
flasgger==0.9.5
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Session==0.4.0
fonttools==4.31.2
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.4.0
kiwisolver==1.4.0
macholib==1.16
MarkupSafe==2.1.1
matplotlib==3.5.1
mistune==2.0.2
multidict==6.0.2
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
PyMySQL==1.0.2
pyobjc==8.1
pyobjc-core==8.1
pyobjc-framework-Accounts==8.1
pyobjc-framework-AddressBook==8.1
pyobjc-framework-AdSupport==8.1
pyobjc-framework-AppleScriptKit==8.1
pyobjc-framework-AppleScriptObjC==8.1
pyobjc-framework-ApplicationServices==8.1
pyobjc-framework-AudioVideoBridging==8.1
pyobjc-framework-Automator==8.1
pyobjc-framework-AVFoundation==8.1
pyobjc-framework-AVKit==8.1
pyobjc-framework-BusinessChat==8.1
pyobjc-framework-CalendarStore==8.1
pyobjc-framework-CFNetwork==8.1
pyobjc-framework-CloudKit==8.1
pyobjc-framework-Cocoa==8.1
pyobjc-framework-Collaboration==8.1
pyobjc-framework-ColorSync==8.1
pyobjc-framework-Contacts==8.1
pyobjc-framework-ContactsUI==8.1
pyobjc-framework-CoreAudio==8.1
pyobjc-framework-CoreAudioKit==8.1
pyobjc-framework-CoreBluetooth==8.1
pyobjc-framework-CoreData==8.1
pyobjc-framework-CoreLocation==8.1
pyobjc-framework-CoreMedia==8.1
pyobjc-framework-CoreMediaIO==8.1
pyobjc-framework-CoreMIDI==8.1
pyobjc-framework-CoreML==8.1
pyobjc-framework-CoreServices==8.1
pyobjc-framework-CoreSpotlight==8.1
pyobjc-framework-CoreText==8.1
pyobjc-framework-CoreWLAN==8.1
pyobjc-framework-CryptoTokenKit==8.1
pyobjc-framework-DictionaryServices==8.1
pyobjc-framework-DiscRecording==8.1
pyobjc-framework-DiscRecordingUI==8.1
pyobjc-framework-DiskArbitration==8.1
pyobjc-framework-DVDPlayback==8.1
pyobjc-framework-EventKit==8.1
pyobjc-framework-ExceptionHandling==8.1
pyobjc-framework-ExternalAccessory==8.1
pyobjc-framework-FinderSync==8.1
pyobjc-framework-FSEvents==8.1
pyobjc-framework-GameCenter==8.1
pyobjc-framework-GameController==8.1
pyobjc-framework-GameKit==8.1
pyobjc-framework-GameplayKit==8.1
pyobjc-framework-ImageCaptureCore==8.1
pyobjc-framework-IMServicePlugIn==8.1
pyobjc-framework-InputMethodKit==8.1
pyobjc-framework-InstallerPlugins==8.1
pyobjc-framework-InstantMessage==8.1
pyobjc-framework-Intents==8.1
pyobjc-framework-IOSurface==8.1
pyobjc-framework-iTunesLibrary==8.1
pyobjc-framework-LatentSemanticMapping==8.1
pyobjc-framework-LaunchServices==8.1
pyobjc-framework-libdispatch==8.1
pyobjc-framework-LocalAuthentication==8.1
pyobjc-framework-MapKit==8.1
pyobjc-framework-MediaAccessibility==8.1
pyobjc-framework-MediaLibrary==8.1
pyobjc-framework-MediaPlayer==8.1
pyobjc-framework-MediaToolbox==8.1
pyobjc-framework-Metal==8.1
pyobjc-framework-MetalKit==8.1
pyobjc-framework-MetalPerformanceShaders==8.1
pyobjc-framework-ModelIO==8.1
pyobjc-framework-MultipeerConnectivity==8.1
pyobjc-framework-NaturalLanguage==8.1
pyobjc-framework-NetFS==8.1
pyobjc-framework-Network==8.1
pyobjc-framework-NetworkExtension==8.1
pyobjc-framework-NotificationCenter==8.1
pyobjc-framework-OpenDirectory==8.1
pyobjc-framework-OSAKit==8.1
pyobjc-framework-Photos==8.1
pyobjc-framework-PhotosUI==8.1
pyobjc-framework-PreferencePanes==8.1
pyobjc-framework-Quartz==8.1
pyobjc-framework-SafariServices==8.1
pyobjc-framework-SceneKit==8.1
pyobjc-framework-ScreenSaver==8.1
pyobjc-framework-ScriptingBridge==8.1
pyobjc-framework-SearchKit==8.1
pyobjc-framework-Security==8.1
pyobjc-framework-SecurityFoundation==8.1
pyobjc-framework-SecurityInterface==8.1
pyobjc-framework-ServiceManagement==8.1
pyobjc-framework-Social==8.1
pyobjc-framework-SpriteKit==8.1
pyobjc-framework-StoreKit==8.1
pyobjc-framework-SyncServices==8.1
pyobjc-framework-SystemConfiguration==8.1
pyobjc-framework-UserNotifications==8.1
pyobjc-framework-VideoSubscriberAccount==8.1
pyobjc-framework-VideoToolbox==8.1
pyobjc-framework-Vision==8.1
pyobjc-framework-WebKit==8.1
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
yarl==1.7.2
